requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch 
torchvision
torchaudio
diffusers
transformers
scipy
safetensors
accelerate
pillow

fastapi 
uvicorn
